requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-filter==2.0.0
mock==2.0.0
Pillow==2.1.0
